refactor(server): attempt to fix next build conf issues
